requirements/base.txt
django==1.3.1
PIL==1.1.7
docutils==0.8.1
pygments==1.4
south==0.7.3
gunicorn==0.13.4
django-grappelli==2.3.5
django-filebrowser==3.4.0
django-extensions==0.7.1
django_compressor==1.1.2
django-debug-toolbar==0.9.4
django-reversion==1.5.1
diff-match-patch
-e hg+http://bitbucket.org/ubernostrum/django-registration/#egg=django-registration